creacion de packeges.ts para interfaz de paquetes
