--- src/app/model/agency.ts
@@ -1,9 +1,0 @@
-export interface Paquete{
-  minimunGroup    :number,
-  categoryId      :number,
-  name	          :string,
-  description     :string,
-  keywords        :string,
-  inversion	      :number,
-  _links?:string
-  }
