+++ src/app/model/packages.ts
@@ -1,0 +1,9 @@
+export interface Paquete{
+    minimunGroup    :number,
+    categoryId      :number,
+    name	          :string,
+    description     :string,
+    keywords        :string,
+    inversion	      :number,
+    _links?:string
+    }
